Show extension on hover
This requires some better logic around unmounting components when a
promise is in-progress.

Once the fetch is being correctly aborted everything will work as
expected.

The copying functionality had to be moved out of the default UI as well,
since it can only be triggered with a user action like click. The
extension will show the download and open controls, but will only copy
the url on click.